feat: adiciona configuração de fuso horário para América/São Paulo e atualiza testes para usar Long como ID

+++ back/src/test/java/repz/app/service/PasswordResetServiceTestIntegration.java
@@ -71,7 +71,7 @@ class PasswordResetServiceTestIntegration extends ServiceIntegrationSupport {
         var updatedToken = tokenRepository.findByToken(token.getToken()).orElseThrow();
         assertThat(updatedToken.getUsed()).isTrue();
 
-        var updatedUser = userRepository.findById(user.getId().intValue()).orElseThrow();
+        var updatedUser = userRepository.findById(Long.valueOf(user.getId())).orElseThrow();
         assertThat(updatedUser.getPassword()).isNotEqualTo("{noop}123456");
     }
